backend/requirements.txt
aiodns==3.0.0
aiofile==3.5.1
aiohttp==3.7.4.post0
asgiref==3.3.4
async-timeout==3.0.1
attrs==21.2.0
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.9.3
bio==0.5.0
biopython==1.79
bs4==0.0.1
caio==0.8.4
certifi==2021.5.30
cffi==1.14.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
distlib==0.3.2
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.4
django-cors-headers==3.7.0
django-crispy-forms==1.12.0
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
djoser==2.1.0
filelock==3.0.12
heroku==0.1.4
idna==2.10
itypes==1.2.0
Jinja2==3.0.1
lxml==4.6.3
MarkupSafe==2.0.1
multidict==5.1.0
numpy==1.19.5
oauthlib==3.1.1
optional-django==0.3.0
Pillow==8.3.1
plac==1.3.3
platformdirs==2.0.2
pycares==4.0.0
pycparser==2.20
pycryptodomex==3.10.1
PyJWT==2.1.0
python-dateutil==1.5
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
react==4.3.0
requests==2.25.1
requests-oauthlib==1.3.0
selenium==3.141.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
soupsieve==2.2.1
sqlparse==0.4.1
static3==0.7.0
stream-chat==3.11.0
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.5
virtualenv==20.5.0
whitenoise==5.2.0
yarl==1.6.3
